Fix remaining SadTabView anti-aliasing issues.
The SadTabView is in a strange situation where it is set
up almost entirely programmatically, but yet it is still
loaded through an xib. Remove the programmatically set up
and arranged elements from the xib entirely, and allocate them
manually.

Use a subclass of NSTextField to with isOpaque overwritten to
enable text anti-aliasing. Use NSTextField instead of NSTextView
because the manual resizing code is written against NSTextField,
and lines get broken in different places when switching to
NSTextView.

BUG=351456

Review URL: https://codereview.chromium.org/202783002