cc: Skip commit state timer on sync compositor
Sync compositor does not use frame rate controller or can have the
circular wait situation mentioned in crbug.com/317430 that caused the
timer to be added. If a wait like that happens, it would be a instant
real deadlock in that case due to all the merged threads.

BUG=344087

Review URL: https://codereview.chromium.org/200793005