components: Eliminate the dependency on chrome's kFrameworkName constant.
Hard coding this string directly on run_all_unittests.cc is the easiest
fix for now and allows us to get rid of this ugly dependency on chrome's
common target.

The right fix should come next with plist file and the
components_test.pak in place, otherwise is not possible to do remove
this dep.

So since alot of infrastructure is needed before we can get the right
fix, this was the minimal change that would allow us to get rid of it
soon rather than later, slowing local and remote builds for no good
reason.

BUG=34856
TEST=components_unittests on Mac works as before
R=blundell@chromium.org,thestig@chromium.org

Review URL: https://codereview.chromium.org/199653005