Remove dead code trying to prevent LayoutTest checkouts
third_party/WebKit/LayoutTests has not been a separate checkout
for years.  src/webkit/data/layout_tests hasn't existed
in any checkout for years.

Remove both of these paths from various DEPS commands on the
masters and slaves as part of auditing all uses of
third_party/WebKit in build/ in preparation for the Blink merge.

This is the master half of the change, the slave half was landed
as https://codereview.chromium.org/177023005

BUG=345464

Review URL: https://codereview.chromium.org/196593004