rAc - fix crash when switching between Wallet accounts on Mac
The Mac view unexpectedly calls delegate::InputIsEditable as a result of view::UpdateAccountChooser, and this causes a crash when wallet_items_ is NULL (as is the case when you've just disabled Wallet).

I believe this bug is present on ToT, but it requires toggling wallet accounts several times to trigger it.

BUG=352726

Review URL: https://codereview.chromium.org/196603023