Set size constraints before native app window creation
This patch is a quick fix for the regression described in the bug.
The window style, which disables maximize of fixed sized windows,
is only set at window creation. Views uses InitParams and a
delegate (which will be NativeAppWindowViews) for the initial
state of a window. The constraints are taken from the delegate.

BUG=351688
TEST=See bug for steps to reproduce and test.

Review URL: https://codereview.chromium.org/198503002