Fix flakyness in ScreenOrientationProviderTests.
There was two causes. The first was that mObserver is initialized
asynchrounously so the tests were flaky because the initial value might
or might not have been the expected one. The second cause was related to
pre-API 17 were we can't reliably listen to 180 degrees orientation
changes.

BUG=353500,353322

Review URL: https://codereview.chromium.org/197213025