Enable immersive fullscreen on Windows Ash.
On ChromeOS a reveal occurs on a swipe down from the top 
edge, and hides on a swipe up from the top edge.

On Windows we can only detect one piece of information: A 
swipe up occurred from the bottom edge, or a swipe down 
occured from the top edge.  These two events are 
indistinguishable on Windows, and they do not come with any 
location information.

As a result, the CrOS behavior of these two distinct 
operations are merged into one operation here.  On Windows 
an "edge swipe" displays the shelf AND does the immersive 
reveal, whereas on CrOS these are separate operations 
triggered by a swipe on the respective edges.

Hiding should still behave the same as CrOS.

BUG=227247

Review URL: https://codereview.chromium.org/198413003