Disable CoreAnimation by default
We're seeing lots of OOM crashes in dev and canary. See if this makes
them go away.

Make sure the TabStripView always gets a layer (regardless of the
flag), since we'll need the layer to use the CoreAnimation tab spinners,
which will be staying in.

BUG=245900

Review URL: https://codereview.chromium.org/203573005