Mojo: Improve the embedder API.
Make CreateChannel() (optionally) take a TaskRunner for the callback.
This avoids having to write thunk code in common cases. (I don't want to
make it automatically grab the TaskRunner for the thread it's called on,
since then you could only call it on a thread with a MessageLoop.)

Also, make it return a (C++) ScopedMessagePipeHandle, instead of a (C)
MojoHandle.

R=sky@chromium.org
TBR=darin@chromium.org

Review URL: https://codereview.chromium.org/203563003