Do null check when setting ScrollClipLayer
WebLayerImpl::setScrollClipLayer() doesn't seem to handle null case
correctly on the layer we passed in.
It calls cc_clip_layer->id() even if cc_clip_layer is set to null.
We want to pass a null clipping layer to stop the layer from scrolling
for the RenderFullscreen, as suggested by aelias in
https://codereview.chromium.org/200943002/

BUG=352183

Review URL: https://codereview.chromium.org/196573035