Cast:Refactor SenderRtpStatistics
This cl removes the need for a designated stats callback,
and shifts the handling of the rtp stats to the audio and video receivers. 
This cl is a first step in adding another callback to the transport which
will return the last transmitted frame.

BUG=353497

Review URL: https://codereview.chromium.org/203263002