Crash fix: WebView does not attempt to alter view hierarchy during WebContents destruction.
The problem is that WebView was attempting to detach the WebContents native view during a call to WebContentsDestroyed().  When the native view (an aura::Window) was detached, the window was hidden, and this in turn caused WebContentsImpl to attempt to background a non-existant render process.

TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/205263002