WebRTC recipe: Fix configuration errors
It turns out the way the configurations are handled in
https://codereview.chromium.org/202163002/ is invalid.
All the configurations for the chromium recipe module
gets the default values instead of the configured ones.

I think this has to do with how I applied apply_config to
the recipe module, which was inspired by the chromium recipe.
Changing it to rely on the automatic handling of the configurations
that happens when api.webrtc.apply_config is called seems to
fix this.

I also took the opportunity to rename the webrtc_standalone config
to just 'webrtc' and add a separate one for ASan.
This made the Mac's get clang=1 set explicitly due to the check in
the default_compiler of the chromium config.

BUG=346198

Review URL: https://codereview.chromium.org/197283037