GN: Split custom into action and action_foreach
Previously the custom target type implemented two modes: iterate over a list of sources, and run a script once with a set of inputs. This proved a bit confusing since to differentiate these cases you had to either fill in sources or not.

This patch splits that apart into two targets: an action for running a script once, and an action_foreach for running a script over a set of inputs. This makes the differentiation explicit and allows more clear error reporting when you mess it up. I also added additional checks.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/200923007