Revert 238561 "Basic autofill into password value on user gestur..."
This CL is for branch 1847.

This is 3rd attempt to do the revert:
 * the original reverting CL https://codereview.chromium.org/203673007/ broke compilation,
 * so it was reverted in https://codereview.chromium.org/203913009
 * fixed CL https://codereview.chromium.org/204463002/ compiled, but broke PasswordAutofillAgentTest.PasswordAutofillTriggersOnChangeEventsOnLoad,
 * so it was reverted in https://codereview.chromium.org/200033003

In addition to the revert itself, the CL also removes part of the test code in chrome/renderer/autofill/password_autofill_agent_browsertest.cc related to r238561, which was added after r238561.

> Basic autofill into password value on user gesture only.
>
> BUG=163072
>
> Review URL: https://codereview.chromium.org/83023017

BUG=337429
R=isherman@chromium.org

Review URL: https://codereview.chromium.org/197883030