Fix crash introduced in https://codereview.chromium.org/179123002/
We could have a dangling pointer if the RenderProcessHost is destroyed
prior to the renderer going away while we have a suspended timer. This
can happen in the case of an intent triggered from an external app since
it reuses one tab, destroying the tab that was previously created by
that app. This happens early during resume, before webkit timers are
resumed.

BUG=352213

Review URL: https://codereview.chromium.org/204933002