Fix a problem that all menu items are always shown.
Although we control the visibility of the menu items, Chrome Shell always shows all menu items.
It is due to setGroupVisible() at the end of ChromeShellActivity.prepareMenu().
setGroupVisible() shows or hides all menu items that are in the given group.
So we need to move setGroupVisible() call to the first line.

Review URL: https://codereview.chromium.org/207753002