Remove webui_test_support target from ui_base.gyp.
This is not right, even if it was, we should not have a separate target
for forwarding ui/ui_resources for generated grit includes. We would
have to be part of ui/resources/ui_resources.gyp:ui_resources target.

What browser_tests was missing were the dependency on ui_resources
target, which generates the webui_resources.h include.

BUG=299841
TEST=browser_tests LINKs.
TBR=ben@chromium.org,thestig@chromium.org

Review URL: https://codereview.chromium.org/207303003