Upload the current device state key on cloud policy registration.
This adds a new field to the registration message in the cloud policy
protocol, providing the current device state key used by the client.
The key can be used by the server to match up the client with a
potentially existing client record on the server stemming from a
previous registration.

BUG=chromium:351663
TEST=unit tests.

Review URL: https://codereview.chromium.org/196053013