aw: Post SynchronousCompositorClient to UI
Do not assume that the callbacks happen on the UI thread, and instead
post them back to UI thread when needed. For accessors, just
return canned value if it involves the view system, or lock protect the
access if it is not

This is a temporary hack until SynchronousCompositor is moved back to
only accessed on UI thread.

BUG=344087

Review URL: https://codereview.chromium.org/203723009