Make possible to check memory allocations inside chromium
This patch implements UncheckedMalloc and UncheckedCalloc for Linux.
Previously it was only possible on Mac. The motivation is to let callers handle
OOM gracefully instead of aborting.

When tcmalloc is used this is implemented via a weak symbol that is overridden
by tcmalloc. This way we get around the problem that neither base cannot depend
on tcmalloc and vica versa. Unfortunately weak symbols are not supported on Windows.
To make that work on Windows one will have to do some build system craft that is similar
of what we do to replace the system malloc with tcmalloc.

This cl does not try to solve the more controversial problem of disallowing
the OOM handler for third party libraries under special circumstances.

BUG=73751

Review URL: https://codereview.chromium.org/55333002