Cast: Enable use of VideoEncodeAccelerator for hardware video encoding
Cast streaming library is currently using GpuVideoAcceleratorFactories
as the interface to access a VideoEncodeAccelerator. This interface
is not actually correct since Content API in Chrome doesn't provide
that. Moreover GpuVideoAcceleratorFactories also covers accelerated
video decoding which is not needed for a cast sender.

This code removes the use of GpuVideoAcceleratorFactories. It is
replaced by two callbacks. One for allocating a VideoEncodeAccelerator
and the second one allocates base::SharedMemory to store output data.

BUG=354267

Review URL: https://codereview.chromium.org/207593002