Mac: Fix for new tab button flashing when dragging a tab out of a window.
When dragging a tab sideways out of a window, the target window is still the
source window for a moment. We should not remove the placeholder and quickly
add it back in this case, because it causes a flash of the new tab button.

BUG=354342

Review URL: https://codereview.chromium.org/208553002