Global cycle detection analysis.
This adds a detect-cycles flag to the plugin which will cause each compilation unit to output a .graph.json file describing the points-to graph of that compilation unit. The process-graph.py script can then be used to process the graph files to detect cycles involving GC roots.

BUG=334149
R=ager@chromium.org

Review URL: https://codereview.chromium.org/207913002