Stop PasswordStoreMac thread in Shutdown().
Currently PasswordStoreMac creates a supplementary thread in Init(), which is called on the UI thread. PasswordStoreMac stops the thread in its destructor. That's a problem if the destructor is not called on the UI thread. The correct way is to stop the thread in Shutdown(), as that's always called on the UI thread.

BUG=350872

Review URL: https://codereview.chromium.org/206503003