Stop keeping track of the current hovered node for accessibility.
For some reason, TalkBack isn't always trying to give accessibility focus
to a node when we send AccessibilityEvent.TYPE_VIEW_HOVER_ENTER on it, so
keep sending it as long as accessibility focus hasn't moved. Just keep track
of whether we're hovering or not as a boolean, instead.

Also, fix the logic in ContentViewCore so that MotionEvent.ACTION_HOVER_EXIT
still gets passed to BrowserAccessibilityManager, just not to the web when
touch exploration is on. We used MotionEvent.ACTION_HOVER_EXIT to know when
the finger is released, to scroll the page.

BUG=353201

Review URL: https://codereview.chromium.org/202283002