Adding functionality to save a machine id to local_state.
The machine id is a 24-bit hash of machine characteristics, stored locally. It is not transmitted. Id generation is platform specific and is currently only implemented on Windows. 

The id will be used to detect when an install of Chrome is copied to a different machine, such as when an install image is deployed across a fleet of machines.

BUG=343273

Review URL: https://codereview.chromium.org/161533002