Add a new telemetry harness uploader bot to chromium.lkgr.
This is an entirely new bot whose only purpose is to package the telemetry
harness and upload it to Google Storage. As the Speed Team only wants these
to be uploaded for LKGR, it makes the most sense to place the bot on
chromium.lkgr.

BUG=349109

Review URL: https://codereview.chromium.org/196403003