[Win] alt + space opens system menu at wrong position in RTL ui.
When window is not maximized system menu opens on the left hand side and its
contents is partialy outside window bounds. After this fix when system menu
is opened using alt + space with RTL locale we draw it on the right hand side
and whole menu is inside browser window.

Review URL: https://codereview.chromium.org/134553002