Fixes password manager action counter and histogram.
An earlier change broke the histogram measurements for actions in the password
manager. This returns a removed (and deprecated) action to an enum so the
counter is correct and adds appropriate deprecation notes to histogram.xml. Also splits kUserActionOverride into two finer grained values, kUserActionOverridePassword and kUserActionOverrideUsernameAndPassword, to give more precise information on the user choice.

BUG=352824

Review URL: https://codereview.chromium.org/196803003