Attempt to fix the flakiness on the win8 builder on the waterfall.
There appear to be cases where in the metro viewer chrome process does not exit causing subsequent
ash tests to fail. One of the errors we have seen in this context is E_APPLICATION_ACTIVATION_TIMED_OUT.

Currently ash_unittests does not clean shutdown the viewer. It closes the channel and terminates the viewer
process via TerminateProcess. That call could fail if the viewer process has a thread in the kernel stack
which could lead to subsequent tests to fail.

Fix is to attempt a clean shutdown of the viewer process. For this we maintain the message filter for the
lifetime of the MetroViewerProcessHost class. We signal an event when the channel is actually closed in
the hosting process which should ensure that the viewer exits cleanly and also wait for the viewer to die. This code
only executes for tests.

BUG=340422
R=cpu@chromium.org, scottmg@chromium.org, cpu

Review URL: https://codereview.chromium.org/203273008