sync: Filter types to apply in configure cycle
Modifies update application so that updates are not applied to unrelated
types during a configure cycle.

This is implemented by limiting the update application step so that it
only affects enabled types.  This also has the side effect of disabling
update application for types while they're throttled, which should have
no effect, since a throttled type is not expected to have any updates to
apply.

This change will allow the non-blocking types to remain enabled while
a configure cycle is in progress.

BUG=351005

Review URL: https://codereview.chromium.org/208263006