Create new Autofill tables as part of DB migration.
Currently, migration from a version 20 database to the latest version fails,
because the migration code assumes that the autofill_dates table exists, but it
was not present at version 20, and is never created during migration.  This CL
fixes the migration path to be more careful and more sane.

BUG=none
TEST=unit_tests
R=pkasting@chromium.org

Review URL: https://codereview.chromium.org/207643002