De-flake AndroidViewIntegrationTest.testReceivingSizeAfterLoadUpdatesLayout
This test was flaking out because loading the test page would produce
either 1 or 2 size change notifications (the first one was the
notification for the about:blank page which is implicitly loaded
before the first navigation).
The fix is to make the test wait for the final, stable size.

BUG=348159
NOTRY=true

Review URL: https://codereview.chromium.org/210033002