Fix tab capture with Mac software compositor again
Make sure that notification messages aren't dropped.

Again, in principle, we should create a new notification to send
indicating a software frame, but, again, this has to be merged to
a long-diverged branch. Also, Mac will soon be on a delegated
renderer and this will all be moot.

BUG=333986

Review URL: https://codereview.chromium.org/199663005