[Hotword] Adding wrapper for audio logging preference to correctly return desired value.
Since the default value for the audio logging preference is "true", just getting the preference would return "true" (duh). But we only want to return true if the user has set the preference -- that is, they have actually opted in. This adds a wrapper around the call to get the value as used in the private api so that we check first if the pref has been set.

BUG=345811

Review URL: https://codereview.chromium.org/208963003