Moved BoundedScrollView out of message_center
Moved BoundedScrollView out of message_center and into views, where it belongs.
Moved all the unit tests as well.

Replaced the BoundedScrollView used in message_center with this one, allowing it
to change the background colour when needed. This also fixes the bug with the
App Info Dialog as the background is not set to a colour specific to the message
center.

BUG=348829

Review URL: https://codereview.chromium.org/193523004