Always list EthernetEAP as a Favorite for debugging
Because EthernetEAP is never a visible service, we do not observe it.
When it is first created, no properties (including Profile) are set, so
it does not appear in the 'Favorite' list and as a result never gets
updated. This currently only affects debugging.

As a workaround, we always treat EthernetEAP networks as a Favorite,
eveb if Profile is empty.

BUG=355117
R=pneubeck@chromium.org

Review URL: https://codereview.chromium.org/210433002