Introduce MediaStreamTrackMetrics class hanging off RTCPeerConnectionHandler.
This takes care of generating lifetime events for tracks sent or
received over a PeerConnection, and supersedes the previous
implementation in RemoteMediaStreamImpl (which I am deleting in this
change).

This change also has minor updates to MediaStreamTrackMetricsHost on
the browser side, to handle lifetime events for both sent and received
tracks and generate UMA histograms of "call duration" for sent and
received tracks.

Improve MockMediaStream to allow multiple observers, otherwise a few
tests would now DCHECK because of the single observer limitation.

Finally, update histograms.xml to add the new histograms for sent tracks.

BUG=348616

Review URL: https://codereview.chromium.org/206873005