Don't dynamically scope GN templates.
Implement defined() for accessors so you can say defined(invoker.bar)

Implement a MakeClosure() function on Scope for cloning and flattening the variables in a scope for later use.

A new Template class that tracks both the template code and the closure information.

Hooking up the new template class to where we use templates.

R=scottmg@chromium.org, scottmg

Review URL: https://codereview.chromium.org/212933008