Cast: Remove LoggingStats in favor of event subscribers.
Removed LoggingStats class and replaced them with StatsEventSubscriber.
StatsEventSubscriber differs in that the type of event to listen to
(audio / video / other) is determined at construction. The API to get
the stats map changed as well - caller now passes in an empty map that
will get assigned with the data. These changes are in line with the
EncodingEventSubscriber. The logic to calculate stats remains the same
as LoggingStats, though this will probably change in the future. Stats
can be reset as well via Reset().

In addition since stats gathering now depends on collection of raw
events (which is enabled everywhere), the flag to enable stats
collection in CastLoggingConfig has been removed.

Removed CastLoggingConfig struct - logging is enabled always now.
There are no use cases where it should be disabled.

BUG=353019

Review URL: https://codereview.chromium.org/210303003