Log call arguments when extensions call the Chrome preferences API.
Without this change, all calls to get or set preferences simply show up
as "types.ChromeSetting.get" or "types.ChromeSetting.set" in the
extension activity log.  With this change, the argument list will
include the name of the preference being changed and (when the
preference is changed) the new value.

BUG=356314

Review URL: https://codereview.chromium.org/214843003