Ensure that extension resources are loaded with the correct scaling applied on Windows.
We were loading the icons with the scaling factor based on the loaded resource pak, which works on
all platforms except on Windows with high dpi enabled with dpi settings other than 100% or 200%.

Fix is to pass in the correct scaling factor when loading these resources like icons etc which ensures that
they get resized correctly.

I made the PlatformGetImageScale function in the ResourceBundle class static and passed in the loaded scale factor
to it. This function returns the correct scale to be used as before. We call this function from the extensions resource
loading code.

There may be other places which need similar fixes. Will address those in upcoming CL's

BUG=351170, 354706
R=sky@chromium.org

Review URL: https://codereview.chromium.org/211493009